bindr: Parametrized Active Bindings

Provides a simple interface for creating active bindings where the bound function accepts additional arguments.

Details

See create_env() for creating an environment populated with active bindings, and populate_env() for populating an existing environment.

See also

Useful links:

https://github.com/krlmlr/bindr

https://krlmlr.github.io/bindr

Report bugs at https://github.com/krlmlr/bindr/issues