bindr-package {bindr}	R Documentation

bindr: Parametrized Active Bindings

Description

Provides a simple interface for creating active bindings where the bound function accepts additional arguments.

Details

See create_env() for creating an environment populated with active bindings, and populate_env() for populating an existing environment.

Author(s)

Maintainer: Kirill Müller krlmlr+r@mailbox.org

Other contributors:

RStudio [copyright holder]

See Also

Useful links:

https://github.com/krlmlr/bindr

https://krlmlr.github.io/bindr

Report bugs at https://github.com/krlmlr/bindr/issues

[Package bindr version 0.1.1 Index]